possibility to draw a false color image immediately after a bvhtracer image, with a dynamically adapting upper limit for the max color

+++ src/main/World.java
@@ -55,7 +55,7 @@ public class World {
 	//public Tracer tracer;
 	public Light ambientLight;
 	public CompositeAABBox bvh;
-	public double maxBVHCounter;
+	public int maxBVHCounter;
 	
 	
 	public World(){
